requirements.txt
tweepy
pandas
pymongo
streamlit
vaderSentiment
python-dotenv
nltk
plotly
langdetect
